refactor: make EAN component reusable
